requirements.txt
# Deepfake Detector - Academic Research Project
# VATASE Radu-Petrut - January 2026

streamlit>=1.30.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy<2,>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
openai>=1.0.0